requirements.txt
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
black==22.10.0
click==8.1.3
decorator==5.1.1
Django==4.1.3
djangorestframework==3.14.0
executing==1.2.0
iniconfig==1.1.1
ipython==8.6.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
traitlets==5.5.0
wcwidth==0.2.5
